Work around javac bug workaround in Gradle
Gradle always passes an internal -XDuseUnsharedTable=true option to
javac to work around a bug in it, which improves compilation speed. This
makes the ExtendJ compiler bail out with an “unknown option” error, and
it's very hard to make Gradle not pass that option. So instead, the
compiler now accepts and silently ignores it.