Lex query and type casing keywords normally
Also changes `do-query` into `query-for`, since the former could
potentially be legal in weirdly written like `do -query while x`.

Since the sequences `query-for` and `switch-type` are *never* legal in
regular Java code, we can actually lex them as regular tokens, instead
of using the hack of parsing them as a keyword, a minus operator and an
identifier.