Store only the path to the KnowBase in Individual
Instead of storing a reference to the KnowBase itself. This makes it
much simpler to serialize Individuals, since now they just consist of
two inert strings. Since the user shouldn't meddle with the KnowBase
class directly, that also puts a little more barrier between them and
getting at an instance of that class.