Intuit ambiguous DL expressions even harder
All ambiguous tokens (⊤, ⊥, ∃, ∀) are now resolved in the lexer by doing
look-ahead into the token stream and returning special type tokens. The
parser only accepts these tokens for types, while it still accepts the
regular tokens both in types and DL expressions. This is enough to
remove the ambiguity in statements like `⊤ ⊔ ⊥ a;`, which are now taken
as a variable declaration, rather than a DL expression followed by a
rogue identifier.